Complete cleanup and refactor. ReadMe still TODO
 On branch jacobSubmission
 Changes to be committed:
	modified:   build.gradle
	modified:   myREADME.md
	modified:   src/main/java/com/varen/alphabet/AlphabetApplication.java
	new file:   src/main/java/com/varen/alphabet/WordSearchService.java
	deleted:    src/main/java/com/varen/alphabet/WorkSearchService.java
	modified:   src/test/java/com/varen/alphabet/AlphabetApplicationTests.java
	deleted:    src/test/resources/sample5.txt
