+++ build.gradle
@@ -2,7 +2,6 @@ plugins {
 	id 'org.springframework.boot' version '2.4.5'
 	id 'io.spring.dependency-management' version '1.0.11.RELEASE'
 	id 'java'
-	id "io.freefair.lombok" version "6.0.0-m2"
 }
 
 group = 'com.varen'
