--- src/test/resources/sample5.txt
@@ -1,5 +1,0 @@
-2x5
-H A S D F
-G E Y B H
-DH
-BYE
